moved nnatba
